:truck: Flatten code structure - move index mappings/definitions
Moved the elastic search document mapping definitions to a more
representative location 'index' module, and flattened the code
structure to not require more packages.

Flat is better than nested.
