--- src/woo_search/search_index/documents/__init__.py
@@ -1,3 +1,0 @@
-from .publications import Document, Publication
-
-__all__ = ["Document", "Publication"]
